requirements.txt
blis==0.7.9
boto3==1.26.10
botocore==1.29.117
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.0.4
cymem==2.0.7
google-pasta==0.2.0
idna==3.4
importlib-metadata==6.5.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
langcodes==3.3.0
MarkupSafe==2.1.2
murmurhash==1.0.9
nltk==3.7
numpy==1.22.4
packaging==23.1
pandas==1.3.5
pathy==0.10.1
Pillow==9.3.0
preshed==3.0.8
protobuf==4.22.3
protobuf3-to-dict==0.1.5
pydantic==1.8.2
python-dateutil==2.8.2
pytz==2023.3
regex==2023.3.23
requests==2.28.2
s3transfer==0.6.0
sagemaker==2.15.0
scikit-learn==1.0.2
scipy==1.10.1
six==1.16.0
smart-open==6.3.0
smdebug-rulesconfig==0.1.5
spacy==3.2.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
thinc==8.0.17
threadpoolctl==3.1.0
torch~=2.0.0
tqdm==4.65.0
training-utils @ git+https://gitlab.com/FaTra/training_utils.git@29728fb3c8828fc14f7dd9dbad7dd5bda71f3bab
typer==0.4.2
typing_extensions==4.5.0
urllib3==1.26.15
wasabi==0.10.1
zipp==3.15.0
setuptools==65.5.0
wandb== 0.12.11